+++ src/main/java/it/beije/quiz/model/User.java
@@ -1,6 +1,5 @@
 package it.beije.quiz.model;
 
-import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonInclude;
 
 import javax.persistence.Column;
@@ -17,7 +16,7 @@ public class User {
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	@Column(name = "id")
-	private Long id;
+	private Integer id;
 	
 	@Column(name = "name")
 	private String name;
@@ -31,10 +30,10 @@ public class User {
 	@Column(name = "password")
 	private String password;
 	
-	public Long getId() {
+	public Integer getId() {
 		return id;
 	}
-	public void setId(Long id) {
+	public void setId(Integer id) {
 		this.id = id;
 	}
